+++ src/compatibility.js
@@ -1,19 +1,14 @@
+import {getColorForDistance} from "./main.js"
+import {highlightTokenShape} from "./util.js"
+
 export function getHexSizeSupportTokenGridCenter(token) {
 	const tokenCenterOffset = CONFIG.hexSizeSupport.getCenterOffset(token)
-	const tokenCenter = {x: token.x + tokenCenterOffset.x, y: token.y + tokenCenterOffset.y}
-	if (token.getFlag("hex-size-support", "borderSize") % 2 === 1)
-		return tokenCenter
-	if (canvas.grid.grid.columns) {
-		let hexOffset = canvas.grid.w / 2
-		if (!CONFIG.hexSizeSupport.getAltOrientationFlag(token))
-			hexOffset *= -1
-		tokenCenter.x += hexOffset
+	return {x: token.x + tokenCenterOffset.x, y: token.y + tokenCenterOffset.y}
+}
+
+export function highlightMeasurementTerrainRuler(ray, startDistance, tokenShape=[{x: 0, y: 0}]) {
+	for (const space of ray.terrainRulerVisitedSpaces.reverse()) {
+		const color = getColorForDistance.call(this, startDistance, space.distance)
+		highlightTokenShape.call(this, space, tokenShape, color)
 	}
-	else {
-		let hexOffset = canvas.grid.h / 2
-		if (CONFIG.hexSizeSupport.getAltOrientationFlag(token))
-			hexOffset *= -1
-		tokenCenter.y += hexOffset
-	}
-	return tokenCenter
 }