+++ src/settings.js
@@ -29,6 +29,15 @@ export function registerSettings() {
 		default: true,
 	})
 
+	game.settings.register(settingsKey, "showGMRulerToPlayers", {
+		name: "drag-ruler.settings.showGMRulerToPlayers.name",
+		hint: "drag-ruler.settings.showGMRulerToPlayers.hint",
+		scope: "world",
+		config: true,
+		type: Boolean,
+		default: true,
+	})
+
 	// This setting will be modified by the api if modules register to it
 	game.settings.register(settingsKey, "speedProvider", {
 		scope: "world",
@@ -111,6 +120,7 @@ class SpeedProviderSettings extends FormApplication {
 	}
 
 	async _updateObject(event, formData) {
+		const selectedSpeedProvider = game.user.isGM ? formData.speedProvider : game.settings.get(settingsKey, "speedProvider")
 		for (let [key, value] of Object.entries(formData)) {
 			// Check if this is color, convert the value to an integer
 			const splitKey = key.split(".", 3)
@@ -121,7 +131,7 @@ class SpeedProviderSettings extends FormApplication {
 			}
 
 			// Don't change settings for speed providers that aren't currently active
-			if (key !== "speedProvider" && !key.startsWith(formData.speedProvider))
+			if (key !== "speedProvider" && !key.startsWith(selectedSpeedProvider))
 				continue
 
 			// Get the key for the current setting